Switch to a snackbar instead of an in canvas notice
