+++ packages/block-library/src/gallery/edit.js
@@ -53,6 +53,11 @@ import useShortCodeTransform from './use-short-code-transform';
 import useGetNewImages from './use-get-new-images';
 import useGetMedia from './use-get-media';
 
+/**
+ * Internal dependencies
+ */
+import useMobileWarning from './use-mobile-warning';
+
 const MAX_COLUMNS = 8;
 const linkOptions = [
 	{ value: LINK_DESTINATION_ATTACHMENT, label: __( 'Attachment Page' ) },
@@ -92,7 +97,7 @@ function GalleryEdit( props ) {
 		shortCodeTransforms,
 		sizeSlug,
 	} = attributes;
-
+	useMobileWarning();
 	const {
 		__unstableMarkNextChangeAsNotPersistent,
 		replaceInnerBlocks,
