+++ packages/block-library/src/gallery/gallery.js
@@ -13,11 +13,6 @@ import { __ } from '@wordpress/i18n';
 import { createBlock } from '@wordpress/blocks';
 import { View } from '@wordpress/primitives';
 
-/**
- * Internal dependencies
- */
-import MobileWarning from './mobile-warning';
-
 const allowedBlocks = [ 'core/image' ];
 
 export const Gallery = ( props ) => {
@@ -73,7 +68,6 @@ export const Gallery = ( props ) => {
 				}
 			) }
 		>
-			<MobileWarning />
 			{ children }
 
 			<View
